Create Nashville Housing SQL Project
This is a SQL project I completed in Microsoft SQL Server with a housing dataset.

The purpose of this project was to use SQL to transform the data into a form that would be more useful for analysis.  This included breaking addresses into multiple columns, deleting unused columns, removing duplicates, and changing to letters Y & N to Yes and No.

This project focused on primary SQL clauses such as SELECT, FROM, WHERE,  and ORDER BY. This project also employed the use of joins and window functions.

These are the queries I ran that  fulfilled each task I needed to accomplish.  The commented-out titles explain the purpose of each set of queries.